:goal_net: add check for deleteTransporter

+++ src/main/java/be/howest/ti/mars/logic/controller/DefaultMarsController.java
@@ -210,6 +210,10 @@ public class DefaultMarsController implements MarsController {
             throw new IllegalArgumentException(AN_EMPTY_ARGUMENT_IS_NOT_ALLOWED);
         }
 
+        if (repository.getTransporter(transporterID) == null){
+            throw new NoSuchElementException("Transporter doesn't exists!");
+        }
+
         Transporter transporter = getTransporter(transporterID);
         repository.deleteTransporter(transporter);
     }
